YDaiReserves condition refactor

+++ contracts/interfaces/IMarket.sol
@@ -2,8 +2,8 @@ pragma solidity ^0.6.10;
 
 
 interface IMarket {
-    function sellDai(address from, address to, uint128 daiIn) external returns(uint256);
-    function buyDai(address from, address to, uint128 daiOut) external returns(uint256);
-    function sellYDai(address from, address to, uint128 yDaiIn) external returns(uint256);
-    function buyYDai(address from, address to, uint128 yDaiOut) external returns(uint256);
+    function sellDai(address from, address to, uint128 daiIn) external returns(uint128);
+    function buyDai(address from, address to, uint128 daiOut) external returns(uint128);
+    function sellYDai(address from, address to, uint128 yDaiIn) external returns(uint128);
+    function buyYDai(address from, address to, uint128 yDaiOut) external returns(uint128);
 }
